Split ZFlash.cpp
